Build separate distributions for each os
The default distribution contains copies of LLVM
for all architectures and OSes. This comes makes
the size of the JAR to about 1GB.

A separate distribution for linux-x86_64 which
excludes jars for windows, macos and linux other
architectures slims the JAR size down
to around 100MB.

Right now, we only have specialized builds for
linux, macos, windows for x86_64. Other arches
can use the generic distributions which bundles
all architectures