Unify Type.ParamRef and Type.Param
This simplifies the code significantly.
We Type.Param is now also comparable by its id.

Previously, we relied on the source location of
a type parameter to be the identity of the type.
This was annoying to do with compiler generated
type parameters.

This opens the door to removing the dependency
from Type to Binder. I'm trying to make sure
that the resolved types don't refer to the AST
nodes directly.