Disallow void returning expressions in defer statements
There's a bug in HIRGen that causes it to generate the alloca instruction for the result of
a defer expression *after* its use. This is a temporary workaround to deal with it.
I don't see a real use case to have defers that return values
so I'll deal with this later. For now, it's better to show
a diagnostic than crashing