[refactor]: Create an immutable Env type
This doesn't do much yet. Eventually, the goal is to be a bit more principled
about the scopes during type checking. The current typechecker has no
concept of "trying" a type equality. For example, when trying to compute
what foo.bar refers to (an extension or a field access or a trait method),
we check a few type equalities in a sequence, but never undo the instantiation
of existential variables if the type doesn't match. This means checking whether
any candidate is applicable leaves garbage type equalities in the environment.